requirements.txt
# AlphaPulse Consolidated Requirements
# Combined requirements for all project components

# =============================================================================
# CORE FRAMEWORK
# =============================================================================
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=11.0.3
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# DATABASE & CACHE
# =============================================================================
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
redis>=5.0.8

# =============================================================================
# DATA PROCESSING & ANALYSIS
# =============================================================================
pandas>=2.2.3
numpy>=1.26.4
scipy>=1.11.4
ta>=0.10.2
yfinance>=0.2.18
ccxt>=4.2.0

# =============================================================================
# MACHINE LEARNING - CORE
# =============================================================================
scikit-learn>=1.6.0
xgboost>=2.1.1
lightgbm>=4.1.0
catboost>=1.2.2
river>=0.20.1

# =============================================================================
# DEEP LEARNING
# =============================================================================
tensorflow>=2.15.0
torch>=2.1.1
torchvision>=0.16.1
torchaudio>=2.1.1
transformers>=4.35.0
tokenizers>=0.14.0
accelerate>=0.24.1
bitsandbytes>=0.41.1

# =============================================================================
# AI & REINFORCEMENT LEARNING
# =============================================================================
stable-baselines3>=2.1.0
gym>=0.26.2

# =============================================================================
# ADVANCED ML FEATURES
# =============================================================================
optuna>=3.4.0
hyperopt>=0.2.7
shap>=0.44.0
lime>=0.2.0.1
eli5>=0.14.0
mlflow>=2.8.1
wandb>=0.16.1
datasets>=2.14.6
evaluate>=0.4.1
peft>=0.7.1

# =============================================================================
# ONNX OPTIMIZATION
# =============================================================================
onnxruntime>=1.22.0
onnxruntime-gpu>=1.22.0
onnx>=1.18.0
skl2onnx>=1.18.0
onnxconverter-common>=1.14.0

# =============================================================================
# SENTIMENT ANALYSIS
# =============================================================================
textblob>=0.17.1
tweepy>=4.14.0
vaderSentiment>=3.3.2

# =============================================================================
# VISUALIZATION & DASHBOARD
# =============================================================================
plotly>=5.24.1
dash>=2.18.1
dash-bootstrap-components>=1.5.0

# =============================================================================
# WEB FRAMEWORK (Alternative)
# =============================================================================
flask>=3.0.3
flask-cors>=5.0.0
gunicorn>=23.0.0

# =============================================================================
# HTTP CLIENTS
# =============================================================================
aiohttp>=3.9.1
httpx>=0.25.2

# =============================================================================
# MONITORING & METRICS
# =============================================================================
prometheus-client>=0.19.0
structlog>=23.2.0
psutil>=5.9.6

# =============================================================================
# SECURITY
# =============================================================================
cryptography>=41.0.8
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0
python-multipart>=0.0.6
click>=8.1.7
rich>=13.7.0
pytz>=2023.3
schedule>=1.2.0

# =============================================================================
# TESTING
# =============================================================================
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# =============================================================================
# DEVELOPMENT TOOLS (Optional for Production)
# =============================================================================
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# =============================================================================
# FRONTEND DEPENDENCIES (Node.js - managed separately)
# =============================================================================
# Note: Frontend dependencies are managed in frontend/package.json
# Run 'npm install' in the frontend directory to install Node.js dependencies
